Extract sections into separate files.
Hopefully much easier to sync with the spec text now.
